Indexed function parameter arguments
For quicker lookups.
When calling function argument, the new environment -expect from the dictionary- contains an array with the EvalObjects of each argument.
Each argument identifier -if it is an argument parameter- holds an index. The index is the argument position. That is, there is no need for lookups in a dictionary.